- Separated the utilities classes into burp, misc, and scan. Will probably change later, but this makes sense right now
