+++ src/burp/model/insertionPoints/ParamNameInsertionPoint.java
@@ -1,8 +1,8 @@
 package burp.model.insertionPoints;
 
 import burp.IParameter;
-import burp.model.utilities.Keysmith;
-import burp.model.utilities.Utilities;
+import burp.model.utilities.misc.Keysmith;
+import burp.model.utilities.misc.Utilities;
 
 import java.util.ArrayList;
 import java.util.HashMap;
