+++ src/burp/model/scanning/guessing/param/ParamAttack.java
@@ -2,16 +2,16 @@ package burp.model.scanning.guessing.param;
 
 import burp.IHttpRequestResponse;
 import burp.IParameter;
-import burp.model.utilities.Keysmith;
-import burp.model.utilities.WordProvider;
-import burp.model.utilities.Attack;
+import burp.model.utilities.misc.Keysmith;
+import burp.model.utilities.misc.WordProvider;
+import burp.model.utilities.scan.Attack;
 import burp.model.insertionPoints.HeaderNameInsertionPoint;
 import burp.model.insertionPoints.JsonParamNameInsertionPoint;
 import burp.model.insertionPoints.ParamNameInsertionPoint;
-import burp.model.utilities.Utilities;
+import burp.model.utilities.misc.Utilities;
 import burp.view.ConfigurableSettings;
 import burp.model.insertionPoints.ParamInsertionPoint;
-import burp.model.utilities.PayloadInjector;
+import burp.model.utilities.scan.PayloadInjector;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonParser;
 import org.apache.commons.collections4.queue.CircularFifoQueue;
