+++ src/burp/model/scanning/guessing/param/ParamGuesser.java
@@ -5,19 +5,19 @@ import burp.IHttpService;
 import burp.IParameter;
 import burp.IResponseVariations;
 import burp.IScannerInsertionPoint;
-import burp.model.utilities.Keysmith;
-import burp.model.utilities.Probe;
-import burp.model.utilities.WordProvider;
-import burp.model.utilities.Attack;
+import burp.model.utilities.misc.Keysmith;
+import burp.model.utilities.scan.Probe;
+import burp.model.utilities.misc.WordProvider;
+import burp.model.utilities.scan.Attack;
 import burp.model.scanning.guessing.header.HeaderMutationGuesser;
 import burp.model.insertionPoints.ParamNameInsertionPoint;
 import burp.model.scanning.BulkScan;
-import burp.model.utilities.Utilities;
+import burp.model.utilities.misc.Utilities;
 import burp.view.ConfigurableSettings;
-import burp.model.utilities.CustomScanIssue;
+import burp.model.utilities.burp.CustomScanIssue;
 import burp.model.insertionPoints.ParamInsertionPoint;
 import burp.model.scanning.ParamScan;
-import burp.model.utilities.PayloadInjector;
+import burp.model.utilities.scan.PayloadInjector;
 import burp.model.insertionPoints.RawInsertionPoint;
 import burp.model.scanning.Scan;
 import org.apache.commons.collections4.queue.CircularFifoQueue;
