+++ src/burp/model/scanning/mining/FatGetScan.java
@@ -5,8 +5,8 @@ import burp.IScanIssue;
 import burp.IScannerInsertionPoint;
 import burp.model.scanning.BulkScanLauncher;
 import burp.model.scanning.ParamScan;
-import burp.model.utilities.Resp;
-import burp.model.utilities.Utilities;
+import burp.model.utilities.scan.Resp;
+import burp.model.utilities.misc.Utilities;
 
 import java.util.List;
