+++ src/burp/model/utilities/burp/CustomScanIssue.java
@@ -1,4 +1,4 @@
-package burp.model.utilities;
+package burp.model.utilities.burp;
 
 import burp.IHttpRequestResponse;
 import burp.IHttpService;
