+++ src/burp/model/utilities/misc/RandomComparator.java
@@ -1,4 +1,4 @@
-package burp.model.utilities;
+package burp.model.utilities.misc;
 
 import java.util.Comparator;
