+++ src/burp/model/utilities/misc/WordProvider.java
@@ -1,4 +1,4 @@
-package burp.model.utilities;
+package burp.model.utilities.misc;
 
 import java.io.File;
 import java.io.FileNotFoundException;
